requirements.txt
python-dotenv
requests
pandas
sentence-transformers
# faiss-cpu     # optional - uncomment on Linux/Kaggle if needed
